Remove go1.12 support (#2679)
* Revert "Adding ppc64le architecture support on travis-ci (#2538)"

This reverts commit fca3f95d7cdfdef203c78f220b84118f44590512.

* not support go1.12

* fix

* Update errors_test.go

* Update debug.go
