+++ .travis.yml
@@ -3,8 +3,6 @@ language: go
 matrix:
   fast_finish: true
   include:
-  - go: 1.12.x
-    env: GO111MODULE=on
   - go: 1.13.x
   - go: 1.13.x
     env:
